+++ tests/transform/transformers/test_image_transformer.py
